fix(playerbot): Change MountSpeed enum from uint8 to uint16
Values 280 and 310 exceed uint8 maximum (255), causing compilation error.
Changed underlying type to uint16 to accommodate all enum values.

+++ src/modules/Playerbot/Companion/MountManager.h
@@ -39,7 +39,7 @@ enum class MountType : uint8
 /**
  * @brief Mount speed categories
  */
-enum class MountSpeed : uint8
+enum class MountSpeed : uint16
 {
     SLOW = 60,           // 60% ground speed
     NORMAL = 100,        // 100% ground speed
